+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Disabled GenAI composer inputs and model/data editors while Generate or Ask requests are active.
+
 ### Changed
 
 ### Removed